+++ bin/run_dwelling.py
@@ -117,6 +117,9 @@ dwelling_args = {
         #     "soc_init": 0.5,
         #     "self_consumption_mode": True,
         # },
+        # "Clothes Dryer": {
+        #     "equipment_class": EventDataLoad,
+        # },
     },
 }