+++ bin/run_equipment.py
@@ -10,6 +10,9 @@ from ochre import (
     Battery,
     ElectricResistanceWaterHeater,
     AirConditioner,
+    ScheduledLoad,
+    EventBasedLoad,
+    EventDataLoad,
 )
 from ochre import CreateFigures
 from ochre.Models.Envelope import Envelope
@@ -321,16 +324,92 @@ def run_hvac():
     CreateFigures.plt.show()
 
 
+def run_scheduled_load(name):
+    # create schedule
+    times = pd.date_range(
+        default_args["start_time"],
+        default_args["start_time"] + default_args["duration"],
+        freq=default_args["time_res"],
+        inclusive="left",
+    )
+    peak_load = 5  # kW
+    power = np.random.choice([0, peak_load], p=[0.98, 0.02], size=len(times))
+    schedule = pd.DataFrame({f"{name} (kW)": power}, index=times)
+
+    equipment_args = {
+        "name": name,
+        "schedule": schedule,
+        **default_args,
+    }
+
+    # Initialize equipment
+    device = ScheduledLoad(**equipment_args)
+
+    # Simulate equipment
+    df = device.simulate()
+
+    print(df.head())
+    df.plot()
+    CreateFigures.plt.show()
+
+
+def run_event_based_load(name):
+    # create event schedule
+    s = default_args["start_time"]
+    d = dt.timedelta(days=1)
+    h = dt.timedelta(hours=1)
+    assert default_args["duration"] >= d * 3
+    event_schedule = pd.DataFrame(
+        {
+            "start_time": [s + h * 10, s + d + h * 14, s + d * 2 + h * 17],
+            "end_time": [s + h * 11, s + d + h * 15, s + d * 2 + h * 17.5],
+            # "power": [1, 2, 1],  # average power per event, in kW, for EventBasedLoad
+            "energy": [1, 2, 0.3],  # energy per event, in kWh, for EventDataLoad
+        }
+    )
+
+    equipment_args = default_args.copy()
+    equipment_args.update(
+        {
+            "name": name,
+            "time_res": dt.timedelta(minutes=1),
+            "duration": dt.timedelta(days=3),
+            "event_schedule": event_schedule,
+            # "equipment_class": EventDataLoad,  # used when running within Dwelling
+        }
+    )
+
+    # Initialize equipment
+    # device = EventBasedLoad(**equipment_args)
+    # Note: event data only available for select equipment
+    #  - Clothes Dryer, Cooking Range
+    device = EventDataLoad(**equipment_args)
+
+    # Simulate equipment
+    df = device.simulate()
+
+    print(df.head())
+    df.plot()
+    CreateFigures.plt.show()
+
+
 if __name__ == "__main__":
     # Extract equipment from a Dwelling model
     # run_equipment_from_house_model("Water Heating")
     # run_equipment_from_house_model("PV")  # Must add PV in run_dwelling.py
 
     # Run equipment without a Dwelling model
-    # run_ev()
+    run_ev()
     # run_pv_with_sam()
     # run_battery_from_schedule()
     # run_battery_self_consumption()
     # run_water_heater()
-    run_water_heater_from_file()
+    # run_water_heater_from_file()
     # run_hvac()
+    # run_scheduled_load("Clothes Dryer")
+    # run_scheduled_load("Cooking Range")
+    # run_scheduled_load("Clothes Washer")
+    # run_event_based_load("Clothes Dryer")
+    # run_event_based_load("Cooking Range")
+    # run_event_based_load("Clothes Washer")
+    # run_equipment_from_house_model()