+++ ochre/Equipment/Equipment.py
@@ -3,17 +3,25 @@ import numpy as np
 
 from ochre import Simulator
 from ochre.utils import OCHREException, load_csv
+from ochre.utils.units import kwh_to_therms
 
 
 class Equipment(Simulator):
-    name = 'Generic Equipment'
-    end_use = 'Other'
+    name = "Generic Equipment"
+    end_use = "Other"
     is_electric = True
     is_gas = False
-    modes = ['On', 'Off']  # On and Off assumed as default modes
-    zone_name = 'Indoor'
+    modes = ["On", "Off"]  # On and Off assumed as default modes
+    zone_name = "Indoor"
 
-    def __init__(self, zone_name=None, envelope_model=None, ext_time_res=None, save_ebm_results=False, **kwargs):
+    def __init__(
+        self,
+        zone_name=None,
+        envelope_model=None,
+        ext_time_res=None,
+        save_ebm_results=False,
+        **kwargs,
+    ):
         """
         Base class for all equipment in a dwelling.
         All equipment must have:
@@ -29,8 +37,15 @@ class Equipment(Simulator):
          - The dwelling schedule (or from a player file)
          - Any other information from the dwelling (passed through house_args)
         """
+        # Update zone based on name. Zone defaults to Indoor
         if zone_name is not None:
             self.zone_name = zone_name
+        elif "Exterior" in self.name:
+            self.zone_name = "Outdoor"
+        elif "Garage" in self.name:
+            self.zone_name = "Garage"
+        elif "Basement" in self.name:
+            self.zone_name = "Foundation"
 
         # If envelope model exists, save the zone
         if envelope_model is not None:
@@ -42,44 +57,54 @@ class Equipment(Simulator):
 
         # General parameters
         self.parameters = self.initialize_parameters(**kwargs)
-        self.results_name = self.end_use if self.end_use not in ['Lighting', 'Other'] else self.name
+        self.results_name = self.end_use if self.end_use not in ["Lighting", "Other"] else self.name
         self.save_ebm_results = save_ebm_results
 
         # Power parameters
         self.electric_kw = 0  # in kW
         self.reactive_kvar = 0  # in kVAR
         self.gas_therms_per_hour = 0  # in therms/hour
-        if 'Zp' in kwargs:
+        if "Zp" in kwargs:
             self.zip_data = (
-                np.array([kwargs['Zp'], kwargs['Ip'], kwargs['Pp']]),  # real ZIP parameters
-                np.array([kwargs['Zq'], kwargs['Iq'], kwargs['Pq']]),  # reactive ZIP parameters
-                np.tan(np.arccos(kwargs['pf']))  # power factor multiplier (+ for capacitative, - for inductive)
+                np.array([kwargs["Zp"], kwargs["Ip"], kwargs["Pp"]]),  # real ZIP parameters
+                np.array([kwargs["Zq"], kwargs["Iq"], kwargs["Pq"]]),  # reactive ZIP parameters
+                # power factor multiplier (+ for capacitative, - for inductive)
+                np.tan(np.arccos(kwargs["pf"])),
             )
         else:
             self.zip_data = None
 
         # Sensible and latent heat parameters
+        # FUTURE: separate convection and radiation, move radiation gains to the surfaces around the zone
         self.sensible_gain = 0  # in W
         self.latent_gain = 0  # in W
+        self.sensible_gain_fraction = kwargs.get("Convective Gain Fraction (-)", 0) + kwargs.get(
+            "Radiative Gain Fraction (-)", 0
+        )
+        self.latent_gain_fraction = kwargs.get("Latent Gain Fraction (-)", 0)
+        assert self.sensible_gain_fraction + self.latent_gain_fraction <= 1.001
 
         # Mode and controller parameters (assuming a duty cycle)
-        self.mode = 'Off'
+        self.mode = "Off"
         self.time_in_mode = dt.timedelta(minutes=0)
         # self.tot_mode_counters = {mode: dt.timedelta(minutes=0) for mode in self.modes}
         self.mode_cycles = {mode: 0 for mode in self.modes}
 
         # Minimum On/Off Times
-        on_time = kwargs.get(self.end_use + ' Minimum On Time', 0)
-        off_time = kwargs.get(self.end_use + ' Minimum Off Time', 0)
+        on_time = kwargs.get(self.end_use + " Minimum On Time", 0)
+        off_time = kwargs.get(self.end_use + " Minimum Off Time", 0)
         self.min_time_in_mode = {mode: dt.timedelta(minutes=on_time) for mode in self.modes}
-        self.min_time_in_mode['Off'] = dt.timedelta(minutes=off_time)
+        self.min_time_in_mode["Off"] = dt.timedelta(minutes=off_time)
 
         self.ext_time_res = ext_time_res
         self.ext_mode_counters = {mode: dt.timedelta(minutes=0) for mode in self.modes}
-        self.duty_cycle_by_mode = {mode: 0 for mode in self.modes}  # fraction of time per mode, should sum to 1
-        self.duty_cycle_by_mode['Off'] = 1
+        # fraction of time per mode, should sum to 1
+        self.duty_cycle_by_mode = {mode: 0 for mode in self.modes}
+        self.duty_cycle_by_mode["Off"] = 1
 
-    def initialize_parameters(self, parameter_file=None, name_col='Name', value_col='Value', **kwargs):
+    def initialize_parameters(
+        self, parameter_file=None, name_col="Name", value_col="Value", **kwargs
+    ):
         if parameter_file is None:
             return {}
 
@@ -99,8 +124,10 @@ class Equipment(Simulator):
         if len(duty_cycles) == len(self.modes) - 1:
             duty_cycles.append(1 - sum(duty_cycles))
         if len(duty_cycles) != len(self.modes):
-            raise OCHREException('Error parsing duty cycles. Expected a list of length equal or 1 less than ' +
-                                     'the number of modes ({}): {}'.format(len(self.modes), duty_cycles))
+            raise OCHREException(
+                "Error parsing duty cycles. Expected a list of length equal or 1 less than "
+                + "the number of modes ({}): {}".format(len(self.modes), duty_cycles)
+            )
 
         self.duty_cycle_by_mode = dict(zip(self.modes, duty_cycles))
 
@@ -115,17 +142,23 @@ class Equipment(Simulator):
         :return: list of mode names in order of priority
         """
         if self.ext_time_res is None:
-            raise OCHREException('External control time resolution is not defined for {}.'.format(self.name))
+            raise OCHREException(
+                "External control time resolution is not defined for {}.".format(self.name)
+            )
         if duty_cycles:
             self.update_duty_cycles(*duty_cycles)
 
-        if (self.current_time - self.start_time) % self.ext_time_res == 0 or \
-                sum(self.ext_mode_counters.values(), dt.timedelta(0)) >= self.ext_time_res:
+        if (self.current_time - self.start_time) % self.ext_time_res == 0 or sum(
+            self.ext_mode_counters.values(), dt.timedelta(0)
+        ) >= self.ext_time_res:
             # reset mode counters
             self.ext_mode_counters = {mode: dt.timedelta(minutes=0) for mode in self.modes}
 
-        modes_with_time = [mode for mode in self.modes
-                           if self.ext_mode_counters[mode] / self.ext_time_res < self.duty_cycle_by_mode[mode]]
+        modes_with_time = [
+            mode
+            for mode in self.modes
+            if self.ext_mode_counters[mode] / self.ext_time_res < self.duty_cycle_by_mode[mode]
+        ]
 
         # move previous mode to top of priority list
         if self.mode in modes_with_time and modes_with_time[0] != self.mode:
@@ -133,15 +166,17 @@ class Equipment(Simulator):
             modes_with_time = [self.mode] + modes_with_time
 
         if not len(modes_with_time):
-            self.warn('No available modes, keeping the current mode. '
-                      'Duty cycles: {}; Time per mode: {}'.format(duty_cycles, self.ext_mode_counters))
+            self.warn(
+                "No available modes, keeping the current mode. "
+                "Duty cycles: {}; Time per mode: {}".format(duty_cycles, self.ext_mode_counters)
+            )
             modes_with_time.append(self.mode)
 
         return modes_with_time
 
     def update_external_control(self, control_signal):
         # Overwrite if external control might exist
-        raise OCHREException('Must define external control algorithm for {}'.format(self.name))
+        raise OCHREException("Must define external control algorithm for {}".format(self.name))
 
     def update_internal_control(self):
         # Returns the equipment mode; can return None if the mode doesn't change
@@ -149,12 +184,19 @@ class Equipment(Simulator):
         raise NotImplementedError()
 
     def calculate_power_and_heat(self):
-        raise NotImplementedError()
+        # Note: only updates heat gains
+        #  - need to calculate power in child classes
+        if self.zone is not None:
+            total_power_w = (
+                self.electric_kw + self.gas_therms_per_hour / kwh_to_therms
+            ) * 1000  # in W
+            self.sensible_gain = total_power_w * self.sensible_gain_fraction
+            self.latent_gain = total_power_w * self.latent_gain_fraction
 
     def add_gains_to_zone(self):
         if self.zone is None:
             return
-        
+
         self.zone.internal_sens_gain += self.sensible_gain
         self.zone.internal_latent_gain += self.latent_gain
 
@@ -167,7 +209,7 @@ class Equipment(Simulator):
             return
         if not self.zip_data:
             return
-        
+
         if v == v0:
             pf_mult = self.zip_data[2]
             self.reactive_kvar = self.electric_kw * pf_mult
@@ -190,16 +232,19 @@ class Equipment(Simulator):
             mode = self.mode
 
         # Get voltage, if disconnected then set mode to off
-        voltage = self.current_schedule.get('Voltage (-)', 1)
+        voltage = self.current_schedule.get("Voltage (-)", 1)
         if voltage == 0:
-            mode = 'Off'
+            mode = "Off"
 
         if mode is None or mode == self.mode:
             self.time_in_mode += self.time_res
         else:
             if mode not in self.modes:
                 raise OCHREException(
-                    "Can't set {} mode to {}. Valid modes are: {}".format(self.name, mode, self.modes))
+                    "Can't set {} mode to {}. Valid modes are: {}".format(
+                        self.name, mode, self.modes
+                    )
+                )
             self.mode = mode
             self.time_in_mode = self.time_res
             self.mode_cycles[self.mode] += 1
@@ -212,7 +257,7 @@ class Equipment(Simulator):
 
         # Add heat gains to zone
         self.add_gains_to_zone()
-        
+
         # Run update for subsimulators (e.g., water tank, battery thermal model)
         super().update_model(heat_data)
 
@@ -224,7 +269,7 @@ class Equipment(Simulator):
         # model definition:
         #  - Continuous time model: E_dot = eta*P - eta_d*P_d - P_b
         #  - Constraints:           E_min <= E <= E_max, 0 <= P <= P_max, E(t=0)=E_0
-        #  - Discharge Constraints: 0 <= P_d <= P_d,max, P * P_d = 0 (i.e. no simultaneous charge/discharge) 
+        #  - Discharge Constraints: 0 <= P_d <= P_d,max, P * P_d = 0 (i.e. no simultaneous charge/discharge)
         # where:
         #  - E = energy state (kWh)
         #  - E_min, E_max = energy state constraints (kWh)
@@ -237,11 +282,11 @@ class Equipment(Simulator):
         #  - P_d,max = maximum power, generating/discharging (kW) (only for devices that can generate power)
         #  - eta_d = efficiency, generating/discharging (-) (only for devices that can generate power)
         return {
-            f'{self.results_name} EBM Energy (kWh)': None,
-            f'{self.results_name} EBM Min Energy (kWh)': 0,
-            f'{self.results_name} EBM Max Energy (kWh)': None,
-            f'{self.results_name} EBM Max Power (kW)': None,
-            f'{self.results_name} EBM Efficiency (-)': 1,
+            f"{self.results_name} EBM Energy (kWh)": None,
+            f"{self.results_name} EBM Min Energy (kWh)": 0,
+            f"{self.results_name} EBM Max Energy (kWh)": None,
+            f"{self.results_name} EBM Max Power (kW)": None,
+            f"{self.results_name} EBM Efficiency (-)": 1,
             # f'{self.results_name} EBM Baseline Power (kW)': 0,
             # f'{self.results_name} EBM Max Discharge Power (kW)': 0,
             # f'{self.results_name} EBM Discharge Efficiency (-)': 1,
@@ -252,16 +297,16 @@ class Equipment(Simulator):
 
         # Note: end use power is included in Dwelling.generate_results
         # Note: individual equipment powers are included in ScheduledLoad.generate_results
-        if self.main_simulator:
+        if self.main_simulator or (self.verbosity >= 6 and self.name != self.end_use):
             if self.is_electric:
-                results[f'{self.results_name} Electric Power (kW)'] = self.electric_kw
+                results[f"{self.results_name} Electric Power (kW)"] = self.electric_kw
                 if self.verbosity >= 8:
-                    results[f'{self.results_name} Reactive Power (kVAR)'] = self.reactive_kvar
+                    results[f"{self.results_name} Reactive Power (kVAR)"] = self.reactive_kvar
             if self.is_gas:
-                results[f'{self.results_name} Gas Power (therms/hour)'] = self.gas_therms_per_hour
+                results[f"{self.results_name} Gas Power (therms/hour)"] = self.gas_therms_per_hour
 
         if self.verbosity >= 7:
-            results[f'{self.results_name} Mode'] = self.mode
+            results[f"{self.results_name} Mode"] = self.mode
 
         return results